Split train_transit.rs into subdirectory modules
Split the 1066-line train_transit.rs into 4 focused files under
train_transit/ directory (all under 500 lines):

- types.rs (158 lines): constants, data structures, TrainTransitState
- state.rs (311 lines): impl TrainTransitState methods, Saveable impl
- systems.rs (288 lines): ECS systems, plugin registration
- tests.rs (294 lines): unit tests
- mod.rs (36 lines): re-exports preserving the public API

Part of #1389
